@@ -5,7 +5,7 @@ import (
 	"github.com/nspcc-dev/neo-go/pkg/interop/storage"
 )
 
-func GetList(ctx storage.Context, key interface{}) [][]byte {
+func GetList(ctx storage.Context, key any) [][]byte {
 	data := storage.Get(ctx, key)
 	if data != nil {
 		return std.Deserialize(data.([]byte)).([][]byte)
@@ -15,7 +15,7 @@ func GetList(ctx storage.Context, key interface{}) [][]byte {
 }
 
 // SetSerialized serializes data and puts it into contract storage.
-func SetSerialized(ctx storage.Context, key interface{}, value interface{}) {
+func SetSerialized(ctx storage.Context, key any, value any) {
 	data := std.Serialize(value)
 	storage.Put(ctx, key, data)
 }
