@@ -26,11 +26,11 @@ const (
 )
 
 // nolint:deadcode,unused
-func _deploy(data interface{}, isUpdate bool) {
+func _deploy(data any, isUpdate bool) {
 	ctx := storage.GetContext()
 
 	if isUpdate {
-		args := data.([]interface{})
+		args := data.([]any)
 		version := args[len(args)-1].(int)
 
 		common.CheckVersion(version)
@@ -101,7 +101,7 @@ func switchToNotary(ctx storage.Context) {
 
 // Update method updates contract source code and manifest. It can be invoked
 // only by committee.
-func Update(script []byte, manifest []byte, data interface{}) {
+func Update(script []byte, manifest []byte, data any) {
 	if !common.HasUpdateAccess() {
 		panic("only committee can update contract")
 	}
@@ -195,7 +195,7 @@ func Version() int {
 	return common.Version
 }
 
-func getUserInfo(ctx storage.Context, key interface{}) UserInfo {
+func getUserInfo(ctx storage.Context, key any) UserInfo {
 	it := storage.Find(ctx, key, storage.KeysOnly|storage.RemovePrefix)
 	pubs := [][]byte{}
 	for iterator.Next(it) {
