@@ -1,4 +1,6 @@
 // Package alphabet contains RPC wrappers for NeoFS Alphabet contract.
+//
+// Code generated by neo-go contract generate-rpcwrapper --manifest <file.json> --out <file.go> [--hash <hash>] [--config <config>]; DO NOT EDIT.
 package alphabet
 
 import (
@@ -17,89 +19,89 @@ import (
 
 // CommonBallot is a contract-specific common.Ballot type used by its methods.
 type CommonBallot struct {
-	ID []byte
+	ID     []byte
 	Voters keys.PublicKeys
 	Height *big.Int
 }
 
 // LedgerBlock is a contract-specific ledger.Block type used by its methods.
 type LedgerBlock struct {
-	Hash util.Uint256
-	Version *big.Int
-	PrevHash util.Uint256
-	MerkleRoot util.Uint256
-	Timestamp *big.Int
-	Nonce *big.Int
-	Index *big.Int
-	NextConsensus util.Uint160
+	Hash               util.Uint256
+	Version            *big.Int
+	PrevHash           util.Uint256
+	MerkleRoot         util.Uint256
+	Timestamp          *big.Int
+	Nonce              *big.Int
+	Index              *big.Int
+	NextConsensus      util.Uint160
 	TransactionsLength *big.Int
 }
 
 // LedgerBlockSR is a contract-specific ledger.BlockSR type used by its methods.
 type LedgerBlockSR struct {
-	Hash util.Uint256
-	Version *big.Int
-	PrevHash util.Uint256
-	MerkleRoot util.Uint256
-	Timestamp *big.Int
-	Nonce *big.Int
-	Index *big.Int
-	NextConsensus util.Uint160
+	Hash               util.Uint256
+	Version            *big.Int
+	PrevHash           util.Uint256
+	MerkleRoot         util.Uint256
+	Timestamp          *big.Int
+	Nonce              *big.Int
+	Index              *big.Int
+	NextConsensus      util.Uint160
 	TransactionsLength *big.Int
-	PrevStateRoot util.Uint256
+	PrevStateRoot      util.Uint256
 }
 
 // LedgerTransaction is a contract-specific ledger.Transaction type used by its methods.
 type LedgerTransaction struct {
-	Hash util.Uint256
-	Version *big.Int
-	Nonce *big.Int
-	Sender util.Uint160
-	SysFee *big.Int
-	NetFee *big.Int
+	Hash            util.Uint256
+	Version         *big.Int
+	Nonce           *big.Int
+	Sender          util.Uint160
+	SysFee          *big.Int
+	NetFee          *big.Int
 	ValidUntilBlock *big.Int
-	Script []byte
+	Script          []byte
 }
 
 // LedgerTransactionSigner is a contract-specific ledger.TransactionSigner type used by its methods.
 type LedgerTransactionSigner struct {
-	Account util.Uint160
-	Scopes *big.Int
+	Account          util.Uint160
+	Scopes           *big.Int
 	AllowedContracts []util.Uint160
-	AllowedGroups keys.PublicKeys
-	Rules []*LedgerWitnessRule
+	AllowedGroups    keys.PublicKeys
+	Rules            []*LedgerWitnessRule
 }
 
 // LedgerWitnessCondition is a contract-specific ledger.WitnessCondition type used by its methods.
 type LedgerWitnessCondition struct {
-	Type *big.Int
+	Type  *big.Int
 	Value any
 }
 
 // LedgerWitnessRule is a contract-specific ledger.WitnessRule type used by its methods.
 type LedgerWitnessRule struct {
-	Action *big.Int
+	Action    *big.Int
 	Condition *LedgerWitnessCondition
 }
 
 // ManagementABI is a contract-specific management.ABI type used by its methods.
 type ManagementABI struct {
 	Methods []*ManagementMethod
-	Events []*ManagementEvent
+	Events  []*ManagementEvent
 }
 
 // ManagementContract is a contract-specific management.Contract type used by its methods.
 type ManagementContract struct {
-	ID *big.Int
+	ID            *big.Int
 	UpdateCounter *big.Int
-	Hash util.Uint160
-	NEF []byte
-	Manifest *ManagementManifest
+	Hash          util.Uint160
+	NEF           []byte
+	Manifest      *ManagementManifest
 }
 
 // ManagementEvent is a contract-specific management.Event type used by its methods.
 type ManagementEvent struct {
-	Name string
+	Name   string
 	Params []*ManagementParameter
 }
 
@@ -111,23 +113,23 @@ type ManagementGroup struct {
 
 // ManagementManifest is a contract-specific management.Manifest type used by its methods.
 type ManagementManifest struct {
-	Name string
-	Groups []*ManagementGroup
-	Features map[string]string
+	Name               string
+	Groups             []*ManagementGroup
+	Features           map[string]string
 	SupportedStandards []string
-	ABI *ManagementABI
-	Permissions []*ManagementPermission
-	Trusts []util.Uint160
-	Extra any
+	ABI                *ManagementABI
+	Permissions        []*ManagementPermission
+	Trusts             []util.Uint160
+	Extra              any
 }
 
 // ManagementMethod is a contract-specific management.Method type used by its methods.
 type ManagementMethod struct {
-	Name string
-	Params []*ManagementParameter
+	Name       string
+	Params     []*ManagementParameter
 	ReturnType *big.Int
-	Offset *big.Int
-	Safe bool
+	Offset     *big.Int
+	Safe       bool
 }
 
 // ManagementParameter is a contract-specific management.Parameter type used by its methods.
@@ -139,7 +141,7 @@ type ManagementParameter struct {
 // ManagementPermission is a contract-specific management.Permission type used by its methods.
 type ManagementPermission struct {
 	Contract util.Uint160
-	Methods []string
+	Methods  []string
 }
 
 // Invoker is used by ContractReader to call various safe methods.
@@ -162,14 +164,14 @@ type Actor interface {
 // ContractReader implements safe contract methods.
 type ContractReader struct {
 	invoker Invoker
-	hash util.Uint160
+	hash    util.Uint160
 }
 
 // Contract implements all contract methods.
 type Contract struct {
 	ContractReader
 	actor Actor
-	hash util.Uint160
+	hash  util.Uint160
 }
 
 // NewReader creates an instance of ContractReader using provided contract hash and the given Invoker.
@@ -291,7 +293,7 @@ func (res *CommonBallot) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.ID, err = arr[index].TryBytes()
@@ -300,14 +302,14 @@ func (res *CommonBallot) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Voters, err = func (item stackitem.Item) (keys.PublicKeys, error) {
+	res.Voters, err = func(item stackitem.Item) (keys.PublicKeys, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make(keys.PublicKeys, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (*keys.PublicKey, error) {
+			res[i], err = func(item stackitem.Item) (*keys.PublicKey, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return nil, err
@@ -317,13 +319,13 @@ func (res *CommonBallot) FromStackItem(item stackitem.Item) error {
 					return nil, err
 				}
 				return k, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Voters: %w", err)
 	}
@@ -360,10 +362,10 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -373,7 +375,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -385,7 +387,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.PrevHash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.PrevHash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -395,13 +397,13 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PrevHash: %w", err)
 	}
 
 	index++
-	res.MerkleRoot, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.MerkleRoot, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -411,7 +413,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field MerkleRoot: %w", err)
 	}
@@ -435,7 +437,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.NextConsensus, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.NextConsensus, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -445,7 +447,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field NextConsensus: %w", err)
 	}
@@ -482,10 +484,10 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -495,7 +497,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -507,7 +509,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.PrevHash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.PrevHash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -517,13 +519,13 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PrevHash: %w", err)
 	}
 
 	index++
-	res.MerkleRoot, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.MerkleRoot, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -533,7 +535,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field MerkleRoot: %w", err)
 	}
@@ -557,7 +559,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.NextConsensus, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.NextConsensus, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -567,7 +569,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field NextConsensus: %w", err)
 	}
@@ -579,7 +581,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.PrevStateRoot, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.PrevStateRoot, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -589,7 +591,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PrevStateRoot: %w", err)
 	}
@@ -620,10 +622,10 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -633,7 +635,7 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -651,7 +653,7 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Sender, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Sender, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -661,7 +663,7 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Sender: %w", err)
 	}
@@ -716,10 +718,10 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Account, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Account, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -729,7 +731,7 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Account: %w", err)
 	}
@@ -741,14 +743,14 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.AllowedContracts, err = func (item stackitem.Item) ([]util.Uint160, error) {
+	res.AllowedContracts, err = func(item stackitem.Item) ([]util.Uint160, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]util.Uint160, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (util.Uint160, error) {
+			res[i], err = func(item stackitem.Item) (util.Uint160, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return util.Uint160{}, err
@@ -758,26 +760,26 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 					return util.Uint160{}, err
 				}
 				return u, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field AllowedContracts: %w", err)
 	}
 
 	index++
-	res.AllowedGroups, err = func (item stackitem.Item) (keys.PublicKeys, error) {
+	res.AllowedGroups, err = func(item stackitem.Item) (keys.PublicKeys, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make(keys.PublicKeys, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (*keys.PublicKey, error) {
+			res[i], err = func(item stackitem.Item) (*keys.PublicKey, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return nil, err
@@ -787,19 +789,19 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 					return nil, err
 				}
 				return k, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field AllowedGroups: %w", err)
 	}
 
 	index++
-	res.Rules, err = func (item stackitem.Item) ([]*LedgerWitnessRule, error) {
+	res.Rules, err = func(item stackitem.Item) ([]*LedgerWitnessRule, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -812,7 +814,7 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Rules: %w", err)
 	}
@@ -843,7 +845,7 @@ func (res *LedgerWitnessCondition) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.Type, err = arr[index].TryInteger()
@@ -883,7 +885,7 @@ func (res *LedgerWitnessRule) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.Action, err = arr[index].TryInteger()
@@ -923,10 +925,10 @@ func (res *ManagementABI) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Methods, err = func (item stackitem.Item) ([]*ManagementMethod, error) {
+	res.Methods, err = func(item stackitem.Item) ([]*ManagementMethod, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -939,13 +941,13 @@ func (res *ManagementABI) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Methods: %w", err)
 	}
 
 	index++
-	res.Events, err = func (item stackitem.Item) ([]*ManagementEvent, error) {
+	res.Events, err = func(item stackitem.Item) ([]*ManagementEvent, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -958,7 +960,7 @@ func (res *ManagementABI) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Events: %w", err)
 	}
@@ -989,7 +991,7 @@ func (res *ManagementContract) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.ID, err = arr[index].TryInteger()
@@ -1004,7 +1006,7 @@ func (res *ManagementContract) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1014,7 +1016,7 @@ func (res *ManagementContract) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -1057,10 +1059,10 @@ func (res *ManagementEvent) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1069,13 +1071,13 @@ func (res *ManagementEvent) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
 
 	index++
-	res.Params, err = func (item stackitem.Item) ([]*ManagementParameter, error) {
+	res.Params, err = func(item stackitem.Item) ([]*ManagementParameter, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1088,7 +1090,7 @@ func (res *ManagementEvent) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Params: %w", err)
 	}
@@ -1119,10 +1121,10 @@ func (res *ManagementGroup) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.PublicKey, err = func (item stackitem.Item) (*keys.PublicKey, error) {
+	res.PublicKey, err = func(item stackitem.Item) (*keys.PublicKey, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return nil, err
@@ -1132,7 +1134,7 @@ func (res *ManagementGroup) FromStackItem(item stackitem.Item) error {
 			return nil, err
 		}
 		return k, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PublicKey: %w", err)
 	}
@@ -1169,10 +1171,10 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1181,13 +1183,13 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
 
 	index++
-	res.Groups, err = func (item stackitem.Item) ([]*ManagementGroup, error) {
+	res.Groups, err = func(item stackitem.Item) ([]*ManagementGroup, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1200,20 +1202,20 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Groups: %w", err)
 	}
 
 	index++
-	res.Features, err = func (item stackitem.Item) (map[string]string, error) {
+	res.Features, err = func(item stackitem.Item) (map[string]string, error) {
 		m, ok := item.Value().([]stackitem.MapElement)
 		if !ok {
 			return nil, fmt.Errorf("%s is not a map", item.Type().String())
 		}
 		res := make(map[string]string)
 		for i := range m {
-			k, err := func (item stackitem.Item) (string, error) {
+			k, err := func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1222,11 +1224,11 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (m[i].Key)
+			}(m[i].Key)
 			if err != nil {
 				return nil, fmt.Errorf("key %d: %w", i, err)
 			}
-			v, err := func (item stackitem.Item) (string, error) {
+			v, err := func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1235,27 +1237,27 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (m[i].Value)
+			}(m[i].Value)
 			if err != nil {
 				return nil, fmt.Errorf("value %d: %w", i, err)
 			}
 			res[k] = v
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Features: %w", err)
 	}
 
 	index++
-	res.SupportedStandards, err = func (item stackitem.Item) ([]string, error) {
+	res.SupportedStandards, err = func(item stackitem.Item) ([]string, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]string, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (string, error) {
+			res[i], err = func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1264,13 +1266,13 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field SupportedStandards: %w", err)
 	}
@@ -1282,7 +1284,7 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Permissions, err = func (item stackitem.Item) ([]*ManagementPermission, error) {
+	res.Permissions, err = func(item stackitem.Item) ([]*ManagementPermission, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1295,20 +1297,20 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Permissions: %w", err)
 	}
 
 	index++
-	res.Trusts, err = func (item stackitem.Item) ([]util.Uint160, error) {
+	res.Trusts, err = func(item stackitem.Item) ([]util.Uint160, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]util.Uint160, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (util.Uint160, error) {
+			res[i], err = func(item stackitem.Item) (util.Uint160, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return util.Uint160{}, err
@@ -1318,13 +1320,13 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return util.Uint160{}, err
 				}
 				return u, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Trusts: %w", err)
 	}
@@ -1361,10 +1363,10 @@ func (res *ManagementMethod) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1373,13 +1375,13 @@ func (res *ManagementMethod) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
 
 	index++
-	res.Params, err = func (item stackitem.Item) ([]*ManagementParameter, error) {
+	res.Params, err = func(item stackitem.Item) ([]*ManagementParameter, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1392,7 +1394,7 @@ func (res *ManagementMethod) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Params: %w", err)
 	}
@@ -1441,10 +1443,10 @@ func (res *ManagementParameter) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1453,7 +1455,7 @@ func (res *ManagementParameter) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
@@ -1490,10 +1492,10 @@ func (res *ManagementPermission) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Contract, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Contract, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1503,20 +1505,20 @@ func (res *ManagementPermission) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Contract: %w", err)
 	}
 
 	index++
-	res.Methods, err = func (item stackitem.Item) ([]string, error) {
+	res.Methods, err = func(item stackitem.Item) ([]string, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]string, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (string, error) {
+			res[i], err = func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1525,13 +1527,13 @@ func (res *ManagementPermission) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Methods: %w", err)
 	}
