@@ -1,4 +1,6 @@
 // Package audit contains RPC wrappers for NeoFS Audit contract.
+//
+// Code generated by neo-go contract generate-rpcwrapper --manifest <file.json> --out <file.go> [--hash <hash>] [--config <config>]; DO NOT EDIT.
 package audit
 
 import (
@@ -18,95 +20,95 @@ import (
 // AuditAuditHeader is a contract-specific audit.AuditHeader type used by its methods.
 type AuditAuditHeader struct {
 	Epoch *big.Int
-	CID []byte
-	From *keys.PublicKey
+	CID   []byte
+	From  *keys.PublicKey
 }
 
 // CommonBallot is a contract-specific common.Ballot type used by its methods.
 type CommonBallot struct {
-	ID []byte
+	ID     []byte
 	Voters keys.PublicKeys
 	Height *big.Int
 }
 
 // LedgerBlock is a contract-specific ledger.Block type used by its methods.
 type LedgerBlock struct {
-	Hash util.Uint256
-	Version *big.Int
-	PrevHash util.Uint256
-	MerkleRoot util.Uint256
-	Timestamp *big.Int
-	Nonce *big.Int
-	Index *big.Int
-	NextConsensus util.Uint160
+	Hash               util.Uint256
+	Version            *big.Int
+	PrevHash           util.Uint256
+	MerkleRoot         util.Uint256
+	Timestamp          *big.Int
+	Nonce              *big.Int
+	Index              *big.Int
+	NextConsensus      util.Uint160
 	TransactionsLength *big.Int
 }
 
 // LedgerBlockSR is a contract-specific ledger.BlockSR type used by its methods.
 type LedgerBlockSR struct {
-	Hash util.Uint256
-	Version *big.Int
-	PrevHash util.Uint256
-	MerkleRoot util.Uint256
-	Timestamp *big.Int
-	Nonce *big.Int
-	Index *big.Int
-	NextConsensus util.Uint160
+	Hash               util.Uint256
+	Version            *big.Int
+	PrevHash           util.Uint256
+	MerkleRoot         util.Uint256
+	Timestamp          *big.Int
+	Nonce              *big.Int
+	Index              *big.Int
+	NextConsensus      util.Uint160
 	TransactionsLength *big.Int
-	PrevStateRoot util.Uint256
+	PrevStateRoot      util.Uint256
 }
 
 // LedgerTransaction is a contract-specific ledger.Transaction type used by its methods.
 type LedgerTransaction struct {
-	Hash util.Uint256
-	Version *big.Int
-	Nonce *big.Int
-	Sender util.Uint160
-	SysFee *big.Int
-	NetFee *big.Int
+	Hash            util.Uint256
+	Version         *big.Int
+	Nonce           *big.Int
+	Sender          util.Uint160
+	SysFee          *big.Int
+	NetFee          *big.Int
 	ValidUntilBlock *big.Int
-	Script []byte
+	Script          []byte
 }
 
 // LedgerTransactionSigner is a contract-specific ledger.TransactionSigner type used by its methods.
 type LedgerTransactionSigner struct {
-	Account util.Uint160
-	Scopes *big.Int
+	Account          util.Uint160
+	Scopes           *big.Int
 	AllowedContracts []util.Uint160
-	AllowedGroups keys.PublicKeys
-	Rules []*LedgerWitnessRule
+	AllowedGroups    keys.PublicKeys
+	Rules            []*LedgerWitnessRule
 }
 
 // LedgerWitnessCondition is a contract-specific ledger.WitnessCondition type used by its methods.
 type LedgerWitnessCondition struct {
-	Type *big.Int
+	Type  *big.Int
 	Value any
 }
 
 // LedgerWitnessRule is a contract-specific ledger.WitnessRule type used by its methods.
 type LedgerWitnessRule struct {
-	Action *big.Int
+	Action    *big.Int
 	Condition *LedgerWitnessCondition
 }
 
 // ManagementABI is a contract-specific management.ABI type used by its methods.
 type ManagementABI struct {
 	Methods []*ManagementMethod
-	Events []*ManagementEvent
+	Events  []*ManagementEvent
 }
 
 // ManagementContract is a contract-specific management.Contract type used by its methods.
 type ManagementContract struct {
-	ID *big.Int
+	ID            *big.Int
 	UpdateCounter *big.Int
-	Hash util.Uint160
-	NEF []byte
-	Manifest *ManagementManifest
+	Hash          util.Uint160
+	NEF           []byte
+	Manifest      *ManagementManifest
 }
 
 // ManagementEvent is a contract-specific management.Event type used by its methods.
 type ManagementEvent struct {
-	Name string
+	Name   string
 	Params []*ManagementParameter
 }
 
@@ -118,23 +120,23 @@ type ManagementGroup struct {
 
 // ManagementManifest is a contract-specific management.Manifest type used by its methods.
 type ManagementManifest struct {
-	Name string
-	Groups []*ManagementGroup
-	Features map[string]string
+	Name               string
+	Groups             []*ManagementGroup
+	Features           map[string]string
 	SupportedStandards []string
-	ABI *ManagementABI
-	Permissions []*ManagementPermission
-	Trusts []util.Uint160
-	Extra any
+	ABI                *ManagementABI
+	Permissions        []*ManagementPermission
+	Trusts             []util.Uint160
+	Extra              any
 }
 
 // ManagementMethod is a contract-specific management.Method type used by its methods.
 type ManagementMethod struct {
-	Name string
-	Params []*ManagementParameter
+	Name       string
+	Params     []*ManagementParameter
 	ReturnType *big.Int
-	Offset *big.Int
-	Safe bool
+	Offset     *big.Int
+	Safe       bool
 }
 
 // ManagementParameter is a contract-specific management.Parameter type used by its methods.
@@ -146,7 +148,7 @@ type ManagementParameter struct {
 // ManagementPermission is a contract-specific management.Permission type used by its methods.
 type ManagementPermission struct {
 	Contract util.Uint160
-	Methods []string
+	Methods  []string
 }
 
 // Invoker is used by ContractReader to call various safe methods.
@@ -169,14 +171,14 @@ type Actor interface {
 // ContractReader implements safe contract methods.
 type ContractReader struct {
 	invoker Invoker
-	hash util.Uint160
+	hash    util.Uint160
 }
 
 // Contract implements all contract methods.
 type Contract struct {
 	ContractReader
 	actor Actor
-	hash util.Uint160
+	hash  util.Uint160
 }
 
 // NewReader creates an instance of ContractReader using provided contract hash and the given Invoker.
@@ -286,7 +288,7 @@ func (res *AuditAuditHeader) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.Epoch, err = arr[index].TryInteger()
@@ -301,7 +303,7 @@ func (res *AuditAuditHeader) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.From, err = func (item stackitem.Item) (*keys.PublicKey, error) {
+	res.From, err = func(item stackitem.Item) (*keys.PublicKey, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return nil, err
@@ -311,7 +313,7 @@ func (res *AuditAuditHeader) FromStackItem(item stackitem.Item) error {
 			return nil, err
 		}
 		return k, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field From: %w", err)
 	}
@@ -342,7 +344,7 @@ func (res *CommonBallot) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.ID, err = arr[index].TryBytes()
@@ -351,14 +353,14 @@ func (res *CommonBallot) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Voters, err = func (item stackitem.Item) (keys.PublicKeys, error) {
+	res.Voters, err = func(item stackitem.Item) (keys.PublicKeys, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make(keys.PublicKeys, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (*keys.PublicKey, error) {
+			res[i], err = func(item stackitem.Item) (*keys.PublicKey, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return nil, err
@@ -368,13 +370,13 @@ func (res *CommonBallot) FromStackItem(item stackitem.Item) error {
 					return nil, err
 				}
 				return k, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Voters: %w", err)
 	}
@@ -411,10 +413,10 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -424,7 +426,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -436,7 +438,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.PrevHash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.PrevHash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -446,13 +448,13 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PrevHash: %w", err)
 	}
 
 	index++
-	res.MerkleRoot, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.MerkleRoot, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -462,7 +464,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field MerkleRoot: %w", err)
 	}
@@ -486,7 +488,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.NextConsensus, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.NextConsensus, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -496,7 +498,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field NextConsensus: %w", err)
 	}
@@ -533,10 +535,10 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -546,7 +548,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -558,7 +560,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.PrevHash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.PrevHash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -568,13 +570,13 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PrevHash: %w", err)
 	}
 
 	index++
-	res.MerkleRoot, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.MerkleRoot, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -584,7 +586,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field MerkleRoot: %w", err)
 	}
@@ -608,7 +610,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.NextConsensus, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.NextConsensus, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -618,7 +620,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field NextConsensus: %w", err)
 	}
@@ -630,7 +632,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.PrevStateRoot, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.PrevStateRoot, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -640,7 +642,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PrevStateRoot: %w", err)
 	}
@@ -671,10 +673,10 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -684,7 +686,7 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -702,7 +704,7 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Sender, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Sender, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -712,7 +714,7 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Sender: %w", err)
 	}
@@ -767,10 +769,10 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Account, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Account, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -780,7 +782,7 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Account: %w", err)
 	}
@@ -792,14 +794,14 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.AllowedContracts, err = func (item stackitem.Item) ([]util.Uint160, error) {
+	res.AllowedContracts, err = func(item stackitem.Item) ([]util.Uint160, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]util.Uint160, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (util.Uint160, error) {
+			res[i], err = func(item stackitem.Item) (util.Uint160, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return util.Uint160{}, err
@@ -809,26 +811,26 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 					return util.Uint160{}, err
 				}
 				return u, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field AllowedContracts: %w", err)
 	}
 
 	index++
-	res.AllowedGroups, err = func (item stackitem.Item) (keys.PublicKeys, error) {
+	res.AllowedGroups, err = func(item stackitem.Item) (keys.PublicKeys, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make(keys.PublicKeys, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (*keys.PublicKey, error) {
+			res[i], err = func(item stackitem.Item) (*keys.PublicKey, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return nil, err
@@ -838,19 +840,19 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 					return nil, err
 				}
 				return k, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field AllowedGroups: %w", err)
 	}
 
 	index++
-	res.Rules, err = func (item stackitem.Item) ([]*LedgerWitnessRule, error) {
+	res.Rules, err = func(item stackitem.Item) ([]*LedgerWitnessRule, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -863,7 +865,7 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Rules: %w", err)
 	}
@@ -894,7 +896,7 @@ func (res *LedgerWitnessCondition) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.Type, err = arr[index].TryInteger()
@@ -934,7 +936,7 @@ func (res *LedgerWitnessRule) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.Action, err = arr[index].TryInteger()
@@ -974,10 +976,10 @@ func (res *ManagementABI) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Methods, err = func (item stackitem.Item) ([]*ManagementMethod, error) {
+	res.Methods, err = func(item stackitem.Item) ([]*ManagementMethod, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -990,13 +992,13 @@ func (res *ManagementABI) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Methods: %w", err)
 	}
 
 	index++
-	res.Events, err = func (item stackitem.Item) ([]*ManagementEvent, error) {
+	res.Events, err = func(item stackitem.Item) ([]*ManagementEvent, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1009,7 +1011,7 @@ func (res *ManagementABI) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Events: %w", err)
 	}
@@ -1040,7 +1042,7 @@ func (res *ManagementContract) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.ID, err = arr[index].TryInteger()
@@ -1055,7 +1057,7 @@ func (res *ManagementContract) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1065,7 +1067,7 @@ func (res *ManagementContract) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -1108,10 +1110,10 @@ func (res *ManagementEvent) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1120,13 +1122,13 @@ func (res *ManagementEvent) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
 
 	index++
-	res.Params, err = func (item stackitem.Item) ([]*ManagementParameter, error) {
+	res.Params, err = func(item stackitem.Item) ([]*ManagementParameter, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1139,7 +1141,7 @@ func (res *ManagementEvent) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Params: %w", err)
 	}
@@ -1170,10 +1172,10 @@ func (res *ManagementGroup) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.PublicKey, err = func (item stackitem.Item) (*keys.PublicKey, error) {
+	res.PublicKey, err = func(item stackitem.Item) (*keys.PublicKey, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return nil, err
@@ -1183,7 +1185,7 @@ func (res *ManagementGroup) FromStackItem(item stackitem.Item) error {
 			return nil, err
 		}
 		return k, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PublicKey: %w", err)
 	}
@@ -1220,10 +1222,10 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1232,13 +1234,13 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
 
 	index++
-	res.Groups, err = func (item stackitem.Item) ([]*ManagementGroup, error) {
+	res.Groups, err = func(item stackitem.Item) ([]*ManagementGroup, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1251,20 +1253,20 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Groups: %w", err)
 	}
 
 	index++
-	res.Features, err = func (item stackitem.Item) (map[string]string, error) {
+	res.Features, err = func(item stackitem.Item) (map[string]string, error) {
 		m, ok := item.Value().([]stackitem.MapElement)
 		if !ok {
 			return nil, fmt.Errorf("%s is not a map", item.Type().String())
 		}
 		res := make(map[string]string)
 		for i := range m {
-			k, err := func (item stackitem.Item) (string, error) {
+			k, err := func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1273,11 +1275,11 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (m[i].Key)
+			}(m[i].Key)
 			if err != nil {
 				return nil, fmt.Errorf("key %d: %w", i, err)
 			}
-			v, err := func (item stackitem.Item) (string, error) {
+			v, err := func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1286,27 +1288,27 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (m[i].Value)
+			}(m[i].Value)
 			if err != nil {
 				return nil, fmt.Errorf("value %d: %w", i, err)
 			}
 			res[k] = v
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Features: %w", err)
 	}
 
 	index++
-	res.SupportedStandards, err = func (item stackitem.Item) ([]string, error) {
+	res.SupportedStandards, err = func(item stackitem.Item) ([]string, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]string, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (string, error) {
+			res[i], err = func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1315,13 +1317,13 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field SupportedStandards: %w", err)
 	}
@@ -1333,7 +1335,7 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Permissions, err = func (item stackitem.Item) ([]*ManagementPermission, error) {
+	res.Permissions, err = func(item stackitem.Item) ([]*ManagementPermission, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1346,20 +1348,20 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Permissions: %w", err)
 	}
 
 	index++
-	res.Trusts, err = func (item stackitem.Item) ([]util.Uint160, error) {
+	res.Trusts, err = func(item stackitem.Item) ([]util.Uint160, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]util.Uint160, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (util.Uint160, error) {
+			res[i], err = func(item stackitem.Item) (util.Uint160, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return util.Uint160{}, err
@@ -1369,13 +1371,13 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return util.Uint160{}, err
 				}
 				return u, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Trusts: %w", err)
 	}
@@ -1412,10 +1414,10 @@ func (res *ManagementMethod) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1424,13 +1426,13 @@ func (res *ManagementMethod) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
 
 	index++
-	res.Params, err = func (item stackitem.Item) ([]*ManagementParameter, error) {
+	res.Params, err = func(item stackitem.Item) ([]*ManagementParameter, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1443,7 +1445,7 @@ func (res *ManagementMethod) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Params: %w", err)
 	}
@@ -1492,10 +1494,10 @@ func (res *ManagementParameter) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1504,7 +1506,7 @@ func (res *ManagementParameter) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
@@ -1541,10 +1543,10 @@ func (res *ManagementPermission) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Contract, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Contract, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1554,20 +1556,20 @@ func (res *ManagementPermission) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Contract: %w", err)
 	}
 
 	index++
-	res.Methods, err = func (item stackitem.Item) ([]string, error) {
+	res.Methods, err = func(item stackitem.Item) ([]string, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]string, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (string, error) {
+			res[i], err = func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1576,13 +1578,13 @@ func (res *ManagementPermission) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Methods: %w", err)
 	}
