@@ -1,4 +1,6 @@
 // Package balance contains RPC wrappers for NeoFS Balance contract.
+//
+// Code generated by neo-go contract generate-rpcwrapper --manifest <file.json> --out <file.go> [--hash <hash>] [--config <config>]; DO NOT EDIT.
 package balance
 
 import (
@@ -19,102 +21,102 @@ import (
 // BalanceAccount is a contract-specific balance.Account type used by its methods.
 type BalanceAccount struct {
 	Balance *big.Int
-	Until *big.Int
-	Parent []byte
+	Until   *big.Int
+	Parent  []byte
 }
 
 // BalanceToken is a contract-specific balance.Token type used by its methods.
 type BalanceToken struct {
-	Symbol string
-	Decimals *big.Int
+	Symbol         string
+	Decimals       *big.Int
 	CirculationKey string
 }
 
 // CommonBallot is a contract-specific common.Ballot type used by its methods.
 type CommonBallot struct {
-	ID []byte
+	ID     []byte
 	Voters keys.PublicKeys
 	Height *big.Int
 }
 
 // LedgerBlock is a contract-specific ledger.Block type used by its methods.
 type LedgerBlock struct {
-	Hash util.Uint256
-	Version *big.Int
-	PrevHash util.Uint256
-	MerkleRoot util.Uint256
-	Timestamp *big.Int
-	Nonce *big.Int
-	Index *big.Int
-	NextConsensus util.Uint160
+	Hash               util.Uint256
+	Version            *big.Int
+	PrevHash           util.Uint256
+	MerkleRoot         util.Uint256
+	Timestamp          *big.Int
+	Nonce              *big.Int
+	Index              *big.Int
+	NextConsensus      util.Uint160
 	TransactionsLength *big.Int
 }
 
 // LedgerBlockSR is a contract-specific ledger.BlockSR type used by its methods.
 type LedgerBlockSR struct {
-	Hash util.Uint256
-	Version *big.Int
-	PrevHash util.Uint256
-	MerkleRoot util.Uint256
-	Timestamp *big.Int
-	Nonce *big.Int
-	Index *big.Int
-	NextConsensus util.Uint160
+	Hash               util.Uint256
+	Version            *big.Int
+	PrevHash           util.Uint256
+	MerkleRoot         util.Uint256
+	Timestamp          *big.Int
+	Nonce              *big.Int
+	Index              *big.Int
+	NextConsensus      util.Uint160
 	TransactionsLength *big.Int
-	PrevStateRoot util.Uint256
+	PrevStateRoot      util.Uint256
 }
 
 // LedgerTransaction is a contract-specific ledger.Transaction type used by its methods.
 type LedgerTransaction struct {
-	Hash util.Uint256
-	Version *big.Int
-	Nonce *big.Int
-	Sender util.Uint160
-	SysFee *big.Int
-	NetFee *big.Int
+	Hash            util.Uint256
+	Version         *big.Int
+	Nonce           *big.Int
+	Sender          util.Uint160
+	SysFee          *big.Int
+	NetFee          *big.Int
 	ValidUntilBlock *big.Int
-	Script []byte
+	Script          []byte
 }
 
 // LedgerTransactionSigner is a contract-specific ledger.TransactionSigner type used by its methods.
 type LedgerTransactionSigner struct {
-	Account util.Uint160
-	Scopes *big.Int
+	Account          util.Uint160
+	Scopes           *big.Int
 	AllowedContracts []util.Uint160
-	AllowedGroups keys.PublicKeys
-	Rules []*LedgerWitnessRule
+	AllowedGroups    keys.PublicKeys
+	Rules            []*LedgerWitnessRule
 }
 
 // LedgerWitnessCondition is a contract-specific ledger.WitnessCondition type used by its methods.
 type LedgerWitnessCondition struct {
-	Type *big.Int
+	Type  *big.Int
 	Value any
 }
 
 // LedgerWitnessRule is a contract-specific ledger.WitnessRule type used by its methods.
 type LedgerWitnessRule struct {
-	Action *big.Int
+	Action    *big.Int
 	Condition *LedgerWitnessCondition
 }
 
 // ManagementABI is a contract-specific management.ABI type used by its methods.
 type ManagementABI struct {
 	Methods []*ManagementMethod
-	Events []*ManagementEvent
+	Events  []*ManagementEvent
 }
 
 // ManagementContract is a contract-specific management.Contract type used by its methods.
 type ManagementContract struct {
-	ID *big.Int
+	ID            *big.Int
 	UpdateCounter *big.Int
-	Hash util.Uint160
-	NEF []byte
-	Manifest *ManagementManifest
+	Hash          util.Uint160
+	NEF           []byte
+	Manifest      *ManagementManifest
 }
 
 // ManagementEvent is a contract-specific management.Event type used by its methods.
 type ManagementEvent struct {
-	Name string
+	Name   string
 	Params []*ManagementParameter
 }
 
@@ -126,23 +128,23 @@ type ManagementGroup struct {
 
 // ManagementManifest is a contract-specific management.Manifest type used by its methods.
 type ManagementManifest struct {
-	Name string
-	Groups []*ManagementGroup
-	Features map[string]string
+	Name               string
+	Groups             []*ManagementGroup
+	Features           map[string]string
 	SupportedStandards []string
-	ABI *ManagementABI
-	Permissions []*ManagementPermission
-	Trusts []util.Uint160
-	Extra any
+	ABI                *ManagementABI
+	Permissions        []*ManagementPermission
+	Trusts             []util.Uint160
+	Extra              any
 }
 
 // ManagementMethod is a contract-specific management.Method type used by its methods.
 type ManagementMethod struct {
-	Name string
-	Params []*ManagementParameter
+	Name       string
+	Params     []*ManagementParameter
 	ReturnType *big.Int
-	Offset *big.Int
-	Safe bool
+	Offset     *big.Int
+	Safe       bool
 }
 
 // ManagementParameter is a contract-specific management.Parameter type used by its methods.
@@ -154,23 +156,23 @@ type ManagementParameter struct {
 // ManagementPermission is a contract-specific management.Permission type used by its methods.
 type ManagementPermission struct {
 	Contract util.Uint160
-	Methods []string
+	Methods  []string
 }
 
 // LockEvent represents "Lock" event emitted by the contract.
 type LockEvent struct {
-	TxID []byte
-	From util.Uint160
-	To util.Uint160
+	TxID   []byte
+	From   util.Uint160
+	To     util.Uint160
 	Amount *big.Int
-	Until *big.Int
+	Until  *big.Int
 }
 
 // TransferXEvent represents "TransferX" event emitted by the contract.
 type TransferXEvent struct {
-	From util.Uint160
-	To util.Uint160
-	Amount *big.Int
+	From    util.Uint160
+	To      util.Uint160
+	Amount  *big.Int
 	Details []byte
 }
 
@@ -197,7 +199,7 @@ type Actor interface {
 type ContractReader struct {
 	nep17.TokenReader
 	invoker Invoker
-	hash util.Uint160
+	hash    util.Uint160
 }
 
 // Contract implements all contract methods.
@@ -205,7 +207,7 @@ type Contract struct {
 	ContractReader
 	nep17.TokenWriter
 	actor Actor
-	hash util.Uint160
+	hash  util.Uint160
 }
 
 // NewReader creates an instance of ContractReader using provided contract hash and the given Invoker.
@@ -379,7 +381,7 @@ func (res *BalanceAccount) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.Balance, err = arr[index].TryInteger()
@@ -425,10 +427,10 @@ func (res *BalanceToken) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Symbol, err = func (item stackitem.Item) (string, error) {
+	res.Symbol, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -437,7 +439,7 @@ func (res *BalanceToken) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Symbol: %w", err)
 	}
@@ -449,7 +451,7 @@ func (res *BalanceToken) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.CirculationKey, err = func (item stackitem.Item) (string, error) {
+	res.CirculationKey, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -458,7 +460,7 @@ func (res *BalanceToken) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field CirculationKey: %w", err)
 	}
@@ -489,7 +491,7 @@ func (res *CommonBallot) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.ID, err = arr[index].TryBytes()
@@ -498,14 +500,14 @@ func (res *CommonBallot) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Voters, err = func (item stackitem.Item) (keys.PublicKeys, error) {
+	res.Voters, err = func(item stackitem.Item) (keys.PublicKeys, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make(keys.PublicKeys, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (*keys.PublicKey, error) {
+			res[i], err = func(item stackitem.Item) (*keys.PublicKey, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return nil, err
@@ -515,13 +517,13 @@ func (res *CommonBallot) FromStackItem(item stackitem.Item) error {
 					return nil, err
 				}
 				return k, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Voters: %w", err)
 	}
@@ -558,10 +560,10 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -571,7 +573,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -583,7 +585,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.PrevHash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.PrevHash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -593,13 +595,13 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PrevHash: %w", err)
 	}
 
 	index++
-	res.MerkleRoot, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.MerkleRoot, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -609,7 +611,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field MerkleRoot: %w", err)
 	}
@@ -633,7 +635,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.NextConsensus, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.NextConsensus, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -643,7 +645,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field NextConsensus: %w", err)
 	}
@@ -680,10 +682,10 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -693,7 +695,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -705,7 +707,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.PrevHash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.PrevHash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -715,13 +717,13 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PrevHash: %w", err)
 	}
 
 	index++
-	res.MerkleRoot, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.MerkleRoot, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -731,7 +733,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field MerkleRoot: %w", err)
 	}
@@ -755,7 +757,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.NextConsensus, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.NextConsensus, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -765,7 +767,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field NextConsensus: %w", err)
 	}
@@ -777,7 +779,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.PrevStateRoot, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.PrevStateRoot, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -787,7 +789,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PrevStateRoot: %w", err)
 	}
@@ -818,10 +820,10 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -831,7 +833,7 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -849,7 +851,7 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Sender, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Sender, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -859,7 +861,7 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Sender: %w", err)
 	}
@@ -914,10 +916,10 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Account, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Account, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -927,7 +929,7 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Account: %w", err)
 	}
@@ -939,14 +941,14 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.AllowedContracts, err = func (item stackitem.Item) ([]util.Uint160, error) {
+	res.AllowedContracts, err = func(item stackitem.Item) ([]util.Uint160, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]util.Uint160, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (util.Uint160, error) {
+			res[i], err = func(item stackitem.Item) (util.Uint160, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return util.Uint160{}, err
@@ -956,26 +958,26 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 					return util.Uint160{}, err
 				}
 				return u, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field AllowedContracts: %w", err)
 	}
 
 	index++
-	res.AllowedGroups, err = func (item stackitem.Item) (keys.PublicKeys, error) {
+	res.AllowedGroups, err = func(item stackitem.Item) (keys.PublicKeys, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make(keys.PublicKeys, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (*keys.PublicKey, error) {
+			res[i], err = func(item stackitem.Item) (*keys.PublicKey, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return nil, err
@@ -985,19 +987,19 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 					return nil, err
 				}
 				return k, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field AllowedGroups: %w", err)
 	}
 
 	index++
-	res.Rules, err = func (item stackitem.Item) ([]*LedgerWitnessRule, error) {
+	res.Rules, err = func(item stackitem.Item) ([]*LedgerWitnessRule, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1010,7 +1012,7 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Rules: %w", err)
 	}
@@ -1041,7 +1043,7 @@ func (res *LedgerWitnessCondition) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.Type, err = arr[index].TryInteger()
@@ -1081,7 +1083,7 @@ func (res *LedgerWitnessRule) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.Action, err = arr[index].TryInteger()
@@ -1121,10 +1123,10 @@ func (res *ManagementABI) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Methods, err = func (item stackitem.Item) ([]*ManagementMethod, error) {
+	res.Methods, err = func(item stackitem.Item) ([]*ManagementMethod, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1137,13 +1139,13 @@ func (res *ManagementABI) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Methods: %w", err)
 	}
 
 	index++
-	res.Events, err = func (item stackitem.Item) ([]*ManagementEvent, error) {
+	res.Events, err = func(item stackitem.Item) ([]*ManagementEvent, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1156,7 +1158,7 @@ func (res *ManagementABI) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Events: %w", err)
 	}
@@ -1187,7 +1189,7 @@ func (res *ManagementContract) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.ID, err = arr[index].TryInteger()
@@ -1202,7 +1204,7 @@ func (res *ManagementContract) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1212,7 +1214,7 @@ func (res *ManagementContract) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -1255,10 +1257,10 @@ func (res *ManagementEvent) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1267,13 +1269,13 @@ func (res *ManagementEvent) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
 
 	index++
-	res.Params, err = func (item stackitem.Item) ([]*ManagementParameter, error) {
+	res.Params, err = func(item stackitem.Item) ([]*ManagementParameter, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1286,7 +1288,7 @@ func (res *ManagementEvent) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Params: %w", err)
 	}
@@ -1317,10 +1319,10 @@ func (res *ManagementGroup) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.PublicKey, err = func (item stackitem.Item) (*keys.PublicKey, error) {
+	res.PublicKey, err = func(item stackitem.Item) (*keys.PublicKey, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return nil, err
@@ -1330,7 +1332,7 @@ func (res *ManagementGroup) FromStackItem(item stackitem.Item) error {
 			return nil, err
 		}
 		return k, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PublicKey: %w", err)
 	}
@@ -1367,10 +1369,10 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1379,13 +1381,13 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
 
 	index++
-	res.Groups, err = func (item stackitem.Item) ([]*ManagementGroup, error) {
+	res.Groups, err = func(item stackitem.Item) ([]*ManagementGroup, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1398,20 +1400,20 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Groups: %w", err)
 	}
 
 	index++
-	res.Features, err = func (item stackitem.Item) (map[string]string, error) {
+	res.Features, err = func(item stackitem.Item) (map[string]string, error) {
 		m, ok := item.Value().([]stackitem.MapElement)
 		if !ok {
 			return nil, fmt.Errorf("%s is not a map", item.Type().String())
 		}
 		res := make(map[string]string)
 		for i := range m {
-			k, err := func (item stackitem.Item) (string, error) {
+			k, err := func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1420,11 +1422,11 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (m[i].Key)
+			}(m[i].Key)
 			if err != nil {
 				return nil, fmt.Errorf("key %d: %w", i, err)
 			}
-			v, err := func (item stackitem.Item) (string, error) {
+			v, err := func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1433,27 +1435,27 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (m[i].Value)
+			}(m[i].Value)
 			if err != nil {
 				return nil, fmt.Errorf("value %d: %w", i, err)
 			}
 			res[k] = v
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Features: %w", err)
 	}
 
 	index++
-	res.SupportedStandards, err = func (item stackitem.Item) ([]string, error) {
+	res.SupportedStandards, err = func(item stackitem.Item) ([]string, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]string, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (string, error) {
+			res[i], err = func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1462,13 +1464,13 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field SupportedStandards: %w", err)
 	}
@@ -1480,7 +1482,7 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Permissions, err = func (item stackitem.Item) ([]*ManagementPermission, error) {
+	res.Permissions, err = func(item stackitem.Item) ([]*ManagementPermission, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1493,20 +1495,20 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Permissions: %w", err)
 	}
 
 	index++
-	res.Trusts, err = func (item stackitem.Item) ([]util.Uint160, error) {
+	res.Trusts, err = func(item stackitem.Item) ([]util.Uint160, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]util.Uint160, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (util.Uint160, error) {
+			res[i], err = func(item stackitem.Item) (util.Uint160, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return util.Uint160{}, err
@@ -1516,13 +1518,13 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return util.Uint160{}, err
 				}
 				return u, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Trusts: %w", err)
 	}
@@ -1559,10 +1561,10 @@ func (res *ManagementMethod) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1571,13 +1573,13 @@ func (res *ManagementMethod) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
 
 	index++
-	res.Params, err = func (item stackitem.Item) ([]*ManagementParameter, error) {
+	res.Params, err = func(item stackitem.Item) ([]*ManagementParameter, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1590,7 +1592,7 @@ func (res *ManagementMethod) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Params: %w", err)
 	}
@@ -1639,10 +1641,10 @@ func (res *ManagementParameter) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1651,7 +1653,7 @@ func (res *ManagementParameter) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
@@ -1688,10 +1690,10 @@ func (res *ManagementPermission) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Contract, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Contract, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1701,20 +1703,20 @@ func (res *ManagementPermission) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Contract: %w", err)
 	}
 
 	index++
-	res.Methods, err = func (item stackitem.Item) ([]string, error) {
+	res.Methods, err = func(item stackitem.Item) ([]string, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]string, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (string, error) {
+			res[i], err = func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1723,13 +1725,13 @@ func (res *ManagementPermission) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Methods: %w", err)
 	}
@@ -1778,7 +1780,7 @@ func (e *LockEvent) FromStackItem(item *stackitem.Array) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	e.TxID, err = arr[index].TryBytes()
@@ -1787,7 +1789,7 @@ func (e *LockEvent) FromStackItem(item *stackitem.Array) error {
 	}
 
 	index++
-	e.From, err = func (item stackitem.Item) (util.Uint160, error) {
+	e.From, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1797,13 +1799,13 @@ func (e *LockEvent) FromStackItem(item *stackitem.Array) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field From: %w", err)
 	}
 
 	index++
-	e.To, err = func (item stackitem.Item) (util.Uint160, error) {
+	e.To, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1813,7 +1815,7 @@ func (e *LockEvent) FromStackItem(item *stackitem.Array) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field To: %w", err)
 	}
@@ -1874,10 +1876,10 @@ func (e *TransferXEvent) FromStackItem(item *stackitem.Array) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	e.From, err = func (item stackitem.Item) (util.Uint160, error) {
+	e.From, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1887,13 +1889,13 @@ func (e *TransferXEvent) FromStackItem(item *stackitem.Array) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field From: %w", err)
 	}
 
 	index++
-	e.To, err = func (item stackitem.Item) (util.Uint160, error) {
+	e.To, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1903,7 +1905,7 @@ func (e *TransferXEvent) FromStackItem(item *stackitem.Array) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field To: %w", err)
 	}
