@@ -1,4 +1,6 @@
 // Package container contains RPC wrappers for NeoFS Container contract.
+//
+// Code generated by neo-go contract generate-rpcwrapper --manifest <file.json> --out <file.go> [--hash <hash>] [--config <config>]; DO NOT EDIT.
 package container
 
 import (
@@ -18,7 +20,7 @@ import (
 
 // CommonBallot is a contract-specific common.Ballot type used by its methods.
 type CommonBallot struct {
-	ID []byte
+	ID     []byte
 	Voters keys.PublicKeys
 	Height *big.Int
 }
@@ -26,14 +28,14 @@ type CommonBallot struct {
 // ContainerContainer is a contract-specific container.Container type used by its methods.
 type ContainerContainer struct {
 	Value []byte
-	Sig []byte
-	Pub *keys.PublicKey
+	Sig   []byte
+	Pub   *keys.PublicKey
 	Token []byte
 }
 
 // ContainerContainerSizes is a contract-specific container.ContainerSizes type used by its methods.
 type ContainerContainerSizes struct {
-	CID []byte
+	CID         []byte
 	Estimations []*ContainerEstimation
 }
 
@@ -46,89 +48,89 @@ type ContainerEstimation struct {
 // ContainerExtendedACL is a contract-specific container.ExtendedACL type used by its methods.
 type ContainerExtendedACL struct {
 	Value []byte
-	Sig []byte
-	Pub *keys.PublicKey
+	Sig   []byte
+	Pub   *keys.PublicKey
 	Token []byte
 }
 
 // LedgerBlock is a contract-specific ledger.Block type used by its methods.
 type LedgerBlock struct {
-	Hash util.Uint256
-	Version *big.Int
-	PrevHash util.Uint256
-	MerkleRoot util.Uint256
-	Timestamp *big.Int
-	Nonce *big.Int
-	Index *big.Int
-	NextConsensus util.Uint160
+	Hash               util.Uint256
+	Version            *big.Int
+	PrevHash           util.Uint256
+	MerkleRoot         util.Uint256
+	Timestamp          *big.Int
+	Nonce              *big.Int
+	Index              *big.Int
+	NextConsensus      util.Uint160
 	TransactionsLength *big.Int
 }
 
 // LedgerBlockSR is a contract-specific ledger.BlockSR type used by its methods.
 type LedgerBlockSR struct {
-	Hash util.Uint256
-	Version *big.Int
-	PrevHash util.Uint256
-	MerkleRoot util.Uint256
-	Timestamp *big.Int
-	Nonce *big.Int
-	Index *big.Int
-	NextConsensus util.Uint160
+	Hash               util.Uint256
+	Version            *big.Int
+	PrevHash           util.Uint256
+	MerkleRoot         util.Uint256
+	Timestamp          *big.Int
+	Nonce              *big.Int
+	Index              *big.Int
+	NextConsensus      util.Uint160
 	TransactionsLength *big.Int
-	PrevStateRoot util.Uint256
+	PrevStateRoot      util.Uint256
 }
 
 // LedgerTransaction is a contract-specific ledger.Transaction type used by its methods.
 type LedgerTransaction struct {
-	Hash util.Uint256
-	Version *big.Int
-	Nonce *big.Int
-	Sender util.Uint160
-	SysFee *big.Int
-	NetFee *big.Int
+	Hash            util.Uint256
+	Version         *big.Int
+	Nonce           *big.Int
+	Sender          util.Uint160
+	SysFee          *big.Int
+	NetFee          *big.Int
 	ValidUntilBlock *big.Int
-	Script []byte
+	Script          []byte
 }
 
 // LedgerTransactionSigner is a contract-specific ledger.TransactionSigner type used by its methods.
 type LedgerTransactionSigner struct {
-	Account util.Uint160
-	Scopes *big.Int
+	Account          util.Uint160
+	Scopes           *big.Int
 	AllowedContracts []util.Uint160
-	AllowedGroups keys.PublicKeys
-	Rules []*LedgerWitnessRule
+	AllowedGroups    keys.PublicKeys
+	Rules            []*LedgerWitnessRule
 }
 
 // LedgerWitnessCondition is a contract-specific ledger.WitnessCondition type used by its methods.
 type LedgerWitnessCondition struct {
-	Type *big.Int
+	Type  *big.Int
 	Value any
 }
 
 // LedgerWitnessRule is a contract-specific ledger.WitnessRule type used by its methods.
 type LedgerWitnessRule struct {
-	Action *big.Int
+	Action    *big.Int
 	Condition *LedgerWitnessCondition
 }
 
 // ManagementABI is a contract-specific management.ABI type used by its methods.
 type ManagementABI struct {
 	Methods []*ManagementMethod
-	Events []*ManagementEvent
+	Events  []*ManagementEvent
 }
 
 // ManagementContract is a contract-specific management.Contract type used by its methods.
 type ManagementContract struct {
-	ID *big.Int
+	ID            *big.Int
 	UpdateCounter *big.Int
-	Hash util.Uint160
-	NEF []byte
-	Manifest *ManagementManifest
+	Hash          util.Uint160
+	NEF           []byte
+	Manifest      *ManagementManifest
 }
 
 // ManagementEvent is a contract-specific management.Event type used by its methods.
 type ManagementEvent struct {
-	Name string
+	Name   string
 	Params []*ManagementParameter
 }
 
@@ -140,23 +142,23 @@ type ManagementGroup struct {
 
 // ManagementManifest is a contract-specific management.Manifest type used by its methods.
 type ManagementManifest struct {
-	Name string
-	Groups []*ManagementGroup
-	Features map[string]string
+	Name               string
+	Groups             []*ManagementGroup
+	Features           map[string]string
 	SupportedStandards []string
-	ABI *ManagementABI
-	Permissions []*ManagementPermission
-	Trusts []util.Uint160
-	Extra any
+	ABI                *ManagementABI
+	Permissions        []*ManagementPermission
+	Trusts             []util.Uint160
+	Extra              any
 }
 
 // ManagementMethod is a contract-specific management.Method type used by its methods.
 type ManagementMethod struct {
-	Name string
-	Params []*ManagementParameter
+	Name       string
+	Params     []*ManagementParameter
 	ReturnType *big.Int
-	Offset *big.Int
-	Safe bool
+	Offset     *big.Int
+	Safe       bool
 }
 
 // ManagementParameter is a contract-specific management.Parameter type used by its methods.
@@ -168,13 +170,13 @@ type ManagementParameter struct {
 // ManagementPermission is a contract-specific management.Permission type used by its methods.
 type ManagementPermission struct {
 	Contract util.Uint160
-	Methods []string
+	Methods  []string
 }
 
 // PutSuccessEvent represents "PutSuccess" event emitted by the contract.
 type PutSuccessEvent struct {
 	ContainerID util.Uint256
-	PublicKey *keys.PublicKey
+	PublicKey   *keys.PublicKey
 }
 
 // DeleteSuccessEvent represents "DeleteSuccess" event emitted by the contract.
@@ -185,7 +187,7 @@ type DeleteSuccessEvent struct {
 // SetEACLSuccessEvent represents "SetEACLSuccess" event emitted by the contract.
 type SetEACLSuccessEvent struct {
 	ContainerID []byte
-	PublicKey *keys.PublicKey
+	PublicKey   *keys.PublicKey
 }
 
 // StartEstimationEvent represents "StartEstimation" event emitted by the contract.
@@ -221,14 +223,14 @@ type Actor interface {
 // ContractReader implements safe contract methods.
 type ContractReader struct {
 	invoker Invoker
-	hash util.Uint160
+	hash    util.Uint160
 }
 
 // Contract implements all contract methods.
 type Contract struct {
 	ContractReader
 	actor Actor
-	hash util.Uint160
+	hash  util.Uint160
 }
 
 // NewReader creates an instance of ContractReader using provided contract hash and the given Invoker.
@@ -549,7 +551,7 @@ func (res *CommonBallot) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.ID, err = arr[index].TryBytes()
@@ -558,14 +560,14 @@ func (res *CommonBallot) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Voters, err = func (item stackitem.Item) (keys.PublicKeys, error) {
+	res.Voters, err = func(item stackitem.Item) (keys.PublicKeys, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make(keys.PublicKeys, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (*keys.PublicKey, error) {
+			res[i], err = func(item stackitem.Item) (*keys.PublicKey, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return nil, err
@@ -575,13 +577,13 @@ func (res *CommonBallot) FromStackItem(item stackitem.Item) error {
 					return nil, err
 				}
 				return k, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Voters: %w", err)
 	}
@@ -618,7 +620,7 @@ func (res *ContainerContainer) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.Value, err = arr[index].TryBytes()
@@ -633,7 +635,7 @@ func (res *ContainerContainer) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Pub, err = func (item stackitem.Item) (*keys.PublicKey, error) {
+	res.Pub, err = func(item stackitem.Item) (*keys.PublicKey, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return nil, err
@@ -643,7 +645,7 @@ func (res *ContainerContainer) FromStackItem(item stackitem.Item) error {
 			return nil, err
 		}
 		return k, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Pub: %w", err)
 	}
@@ -680,7 +682,7 @@ func (res *ContainerContainerSizes) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.CID, err = arr[index].TryBytes()
@@ -689,7 +691,7 @@ func (res *ContainerContainerSizes) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Estimations, err = func (item stackitem.Item) ([]*ContainerEstimation, error) {
+	res.Estimations, err = func(item stackitem.Item) ([]*ContainerEstimation, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -702,7 +704,7 @@ func (res *ContainerContainerSizes) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Estimations: %w", err)
 	}
@@ -733,10 +735,10 @@ func (res *ContainerEstimation) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.From, err = func (item stackitem.Item) (*keys.PublicKey, error) {
+	res.From, err = func(item stackitem.Item) (*keys.PublicKey, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return nil, err
@@ -746,7 +748,7 @@ func (res *ContainerEstimation) FromStackItem(item stackitem.Item) error {
 			return nil, err
 		}
 		return k, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field From: %w", err)
 	}
@@ -783,7 +785,7 @@ func (res *ContainerExtendedACL) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.Value, err = arr[index].TryBytes()
@@ -798,7 +800,7 @@ func (res *ContainerExtendedACL) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Pub, err = func (item stackitem.Item) (*keys.PublicKey, error) {
+	res.Pub, err = func(item stackitem.Item) (*keys.PublicKey, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return nil, err
@@ -808,7 +810,7 @@ func (res *ContainerExtendedACL) FromStackItem(item stackitem.Item) error {
 			return nil, err
 		}
 		return k, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Pub: %w", err)
 	}
@@ -845,10 +847,10 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -858,7 +860,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -870,7 +872,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.PrevHash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.PrevHash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -880,13 +882,13 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PrevHash: %w", err)
 	}
 
 	index++
-	res.MerkleRoot, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.MerkleRoot, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -896,7 +898,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field MerkleRoot: %w", err)
 	}
@@ -920,7 +922,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.NextConsensus, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.NextConsensus, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -930,7 +932,7 @@ func (res *LedgerBlock) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field NextConsensus: %w", err)
 	}
@@ -967,10 +969,10 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -980,7 +982,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -992,7 +994,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.PrevHash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.PrevHash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -1002,13 +1004,13 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PrevHash: %w", err)
 	}
 
 	index++
-	res.MerkleRoot, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.MerkleRoot, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -1018,7 +1020,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field MerkleRoot: %w", err)
 	}
@@ -1042,7 +1044,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.NextConsensus, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.NextConsensus, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1052,7 +1054,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field NextConsensus: %w", err)
 	}
@@ -1064,7 +1066,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.PrevStateRoot, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.PrevStateRoot, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -1074,7 +1076,7 @@ func (res *LedgerBlockSR) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PrevStateRoot: %w", err)
 	}
@@ -1105,10 +1107,10 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint256, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -1118,7 +1120,7 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -1136,7 +1138,7 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Sender, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Sender, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1146,7 +1148,7 @@ func (res *LedgerTransaction) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Sender: %w", err)
 	}
@@ -1201,10 +1203,10 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Account, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Account, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1214,7 +1216,7 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Account: %w", err)
 	}
@@ -1226,14 +1228,14 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.AllowedContracts, err = func (item stackitem.Item) ([]util.Uint160, error) {
+	res.AllowedContracts, err = func(item stackitem.Item) ([]util.Uint160, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]util.Uint160, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (util.Uint160, error) {
+			res[i], err = func(item stackitem.Item) (util.Uint160, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return util.Uint160{}, err
@@ -1243,26 +1245,26 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 					return util.Uint160{}, err
 				}
 				return u, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field AllowedContracts: %w", err)
 	}
 
 	index++
-	res.AllowedGroups, err = func (item stackitem.Item) (keys.PublicKeys, error) {
+	res.AllowedGroups, err = func(item stackitem.Item) (keys.PublicKeys, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make(keys.PublicKeys, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (*keys.PublicKey, error) {
+			res[i], err = func(item stackitem.Item) (*keys.PublicKey, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return nil, err
@@ -1272,19 +1274,19 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 					return nil, err
 				}
 				return k, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field AllowedGroups: %w", err)
 	}
 
 	index++
-	res.Rules, err = func (item stackitem.Item) ([]*LedgerWitnessRule, error) {
+	res.Rules, err = func(item stackitem.Item) ([]*LedgerWitnessRule, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1297,7 +1299,7 @@ func (res *LedgerTransactionSigner) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Rules: %w", err)
 	}
@@ -1328,7 +1330,7 @@ func (res *LedgerWitnessCondition) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.Type, err = arr[index].TryInteger()
@@ -1368,7 +1370,7 @@ func (res *LedgerWitnessRule) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.Action, err = arr[index].TryInteger()
@@ -1408,10 +1410,10 @@ func (res *ManagementABI) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Methods, err = func (item stackitem.Item) ([]*ManagementMethod, error) {
+	res.Methods, err = func(item stackitem.Item) ([]*ManagementMethod, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1424,13 +1426,13 @@ func (res *ManagementABI) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Methods: %w", err)
 	}
 
 	index++
-	res.Events, err = func (item stackitem.Item) ([]*ManagementEvent, error) {
+	res.Events, err = func(item stackitem.Item) ([]*ManagementEvent, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1443,7 +1445,7 @@ func (res *ManagementABI) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Events: %w", err)
 	}
@@ -1474,7 +1476,7 @@ func (res *ManagementContract) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	res.ID, err = arr[index].TryInteger()
@@ -1489,7 +1491,7 @@ func (res *ManagementContract) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Hash, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Hash, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1499,7 +1501,7 @@ func (res *ManagementContract) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Hash: %w", err)
 	}
@@ -1542,10 +1544,10 @@ func (res *ManagementEvent) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1554,13 +1556,13 @@ func (res *ManagementEvent) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
 
 	index++
-	res.Params, err = func (item stackitem.Item) ([]*ManagementParameter, error) {
+	res.Params, err = func(item stackitem.Item) ([]*ManagementParameter, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1573,7 +1575,7 @@ func (res *ManagementEvent) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Params: %w", err)
 	}
@@ -1604,10 +1606,10 @@ func (res *ManagementGroup) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.PublicKey, err = func (item stackitem.Item) (*keys.PublicKey, error) {
+	res.PublicKey, err = func(item stackitem.Item) (*keys.PublicKey, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return nil, err
@@ -1617,7 +1619,7 @@ func (res *ManagementGroup) FromStackItem(item stackitem.Item) error {
 			return nil, err
 		}
 		return k, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PublicKey: %w", err)
 	}
@@ -1654,10 +1656,10 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1666,13 +1668,13 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
 
 	index++
-	res.Groups, err = func (item stackitem.Item) ([]*ManagementGroup, error) {
+	res.Groups, err = func(item stackitem.Item) ([]*ManagementGroup, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1685,20 +1687,20 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Groups: %w", err)
 	}
 
 	index++
-	res.Features, err = func (item stackitem.Item) (map[string]string, error) {
+	res.Features, err = func(item stackitem.Item) (map[string]string, error) {
 		m, ok := item.Value().([]stackitem.MapElement)
 		if !ok {
 			return nil, fmt.Errorf("%s is not a map", item.Type().String())
 		}
 		res := make(map[string]string)
 		for i := range m {
-			k, err := func (item stackitem.Item) (string, error) {
+			k, err := func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1707,11 +1709,11 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (m[i].Key)
+			}(m[i].Key)
 			if err != nil {
 				return nil, fmt.Errorf("key %d: %w", i, err)
 			}
-			v, err := func (item stackitem.Item) (string, error) {
+			v, err := func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1720,27 +1722,27 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (m[i].Value)
+			}(m[i].Value)
 			if err != nil {
 				return nil, fmt.Errorf("value %d: %w", i, err)
 			}
 			res[k] = v
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Features: %w", err)
 	}
 
 	index++
-	res.SupportedStandards, err = func (item stackitem.Item) ([]string, error) {
+	res.SupportedStandards, err = func(item stackitem.Item) ([]string, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]string, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (string, error) {
+			res[i], err = func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -1749,13 +1751,13 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field SupportedStandards: %w", err)
 	}
@@ -1767,7 +1769,7 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 	}
 
 	index++
-	res.Permissions, err = func (item stackitem.Item) ([]*ManagementPermission, error) {
+	res.Permissions, err = func(item stackitem.Item) ([]*ManagementPermission, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1780,20 +1782,20 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Permissions: %w", err)
 	}
 
 	index++
-	res.Trusts, err = func (item stackitem.Item) ([]util.Uint160, error) {
+	res.Trusts, err = func(item stackitem.Item) ([]util.Uint160, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]util.Uint160, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (util.Uint160, error) {
+			res[i], err = func(item stackitem.Item) (util.Uint160, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return util.Uint160{}, err
@@ -1803,13 +1805,13 @@ func (res *ManagementManifest) FromStackItem(item stackitem.Item) error {
 					return util.Uint160{}, err
 				}
 				return u, nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Trusts: %w", err)
 	}
@@ -1846,10 +1848,10 @@ func (res *ManagementMethod) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1858,13 +1860,13 @@ func (res *ManagementMethod) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
 
 	index++
-	res.Params, err = func (item stackitem.Item) ([]*ManagementParameter, error) {
+	res.Params, err = func(item stackitem.Item) ([]*ManagementParameter, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
@@ -1877,7 +1879,7 @@ func (res *ManagementMethod) FromStackItem(item stackitem.Item) error {
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Params: %w", err)
 	}
@@ -1926,10 +1928,10 @@ func (res *ManagementParameter) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Name, err = func (item stackitem.Item) (string, error) {
+	res.Name, err = func(item stackitem.Item) (string, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return "", err
@@ -1938,7 +1940,7 @@ func (res *ManagementParameter) FromStackItem(item stackitem.Item) error {
 			return "", errors.New("not a UTF-8 string")
 		}
 		return string(b), nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Name: %w", err)
 	}
@@ -1975,10 +1977,10 @@ func (res *ManagementPermission) FromStackItem(item stackitem.Item) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	res.Contract, err = func (item stackitem.Item) (util.Uint160, error) {
+	res.Contract, err = func(item stackitem.Item) (util.Uint160, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint160{}, err
@@ -1988,20 +1990,20 @@ func (res *ManagementPermission) FromStackItem(item stackitem.Item) error {
 			return util.Uint160{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Contract: %w", err)
 	}
 
 	index++
-	res.Methods, err = func (item stackitem.Item) ([]string, error) {
+	res.Methods, err = func(item stackitem.Item) ([]string, error) {
 		arr, ok := item.Value().([]stackitem.Item)
 		if !ok {
 			return nil, errors.New("not an array")
 		}
 		res := make([]string, len(arr))
 		for i := range res {
-			res[i], err = func (item stackitem.Item) (string, error) {
+			res[i], err = func(item stackitem.Item) (string, error) {
 				b, err := item.TryBytes()
 				if err != nil {
 					return "", err
@@ -2010,13 +2012,13 @@ func (res *ManagementPermission) FromStackItem(item stackitem.Item) error {
 					return "", errors.New("not a UTF-8 string")
 				}
 				return string(b), nil
-			} (arr[i])
+			}(arr[i])
 			if err != nil {
 				return nil, fmt.Errorf("item %d: %w", i, err)
 			}
 		}
 		return res, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field Methods: %w", err)
 	}
@@ -2065,10 +2067,10 @@ func (e *PutSuccessEvent) FromStackItem(item *stackitem.Array) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
-	e.ContainerID, err = func (item stackitem.Item) (util.Uint256, error) {
+	e.ContainerID, err = func(item stackitem.Item) (util.Uint256, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return util.Uint256{}, err
@@ -2078,13 +2080,13 @@ func (e *PutSuccessEvent) FromStackItem(item *stackitem.Array) error {
 			return util.Uint256{}, err
 		}
 		return u, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field ContainerID: %w", err)
 	}
 
 	index++
-	e.PublicKey, err = func (item stackitem.Item) (*keys.PublicKey, error) {
+	e.PublicKey, err = func(item stackitem.Item) (*keys.PublicKey, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return nil, err
@@ -2094,7 +2096,7 @@ func (e *PutSuccessEvent) FromStackItem(item *stackitem.Array) error {
 			return nil, err
 		}
 		return k, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PublicKey: %w", err)
 	}
@@ -2143,7 +2145,7 @@ func (e *DeleteSuccessEvent) FromStackItem(item *stackitem.Array) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	e.ContainerID, err = arr[index].TryBytes()
@@ -2195,7 +2197,7 @@ func (e *SetEACLSuccessEvent) FromStackItem(item *stackitem.Array) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	e.ContainerID, err = arr[index].TryBytes()
@@ -2204,7 +2206,7 @@ func (e *SetEACLSuccessEvent) FromStackItem(item *stackitem.Array) error {
 	}
 
 	index++
-	e.PublicKey, err = func (item stackitem.Item) (*keys.PublicKey, error) {
+	e.PublicKey, err = func(item stackitem.Item) (*keys.PublicKey, error) {
 		b, err := item.TryBytes()
 		if err != nil {
 			return nil, err
@@ -2214,7 +2216,7 @@ func (e *SetEACLSuccessEvent) FromStackItem(item *stackitem.Array) error {
 			return nil, err
 		}
 		return k, nil
-	} (arr[index])
+	}(arr[index])
 	if err != nil {
 		return fmt.Errorf("field PublicKey: %w", err)
 	}
@@ -2263,7 +2265,7 @@ func (e *StartEstimationEvent) FromStackItem(item *stackitem.Array) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	e.Epoch, err = arr[index].TryInteger()
@@ -2315,7 +2317,7 @@ func (e *StopEstimationEvent) FromStackItem(item *stackitem.Array) error {
 
 	var (
 		index = -1
-		err error
+		err   error
 	)
 	index++
 	e.Epoch, err = arr[index].TryInteger()
